+++ swap/src/asb/command.rs
@@ -74,6 +74,15 @@ where
             env_config: env_config(testnet),
             cmd: Command::Config,
         },
+        RawCommand::ExportBitcoinWallet => Arguments {
+            testnet,
+            json,
+            sled,
+            disable_timestamp,
+            config_path: config_path(config, testnet)?,
+            env_config: env_config(testnet),
+            cmd: Command::ExportBitcoinWallet,
+        },
         RawCommand::ManualRecovery(ManualRecovery::Redeem {
             redeem_params: RecoverCommandParams { swap_id },
             do_not_await_finality,
@@ -222,6 +231,7 @@ pub enum Command {
     SafelyAbort {
         swap_id: Uuid,
     },
+    ExportBitcoinWallet,
 }
 
 #[derive(structopt::StructOpt, Debug)]
@@ -296,6 +306,8 @@ pub enum RawCommand {
         about = "Prints the Bitcoin and Monero balance. Requires the monero-wallet-rpc to be running."
     )]
     Balance,
+    #[structopt(about = "Print the internal bitcoin wallet descriptor.")]
+    ExportBitcoinWallet,
     #[structopt(about = "Contains sub-commands for recovering a swap manually.")]
     ManualRecovery(ManualRecovery),
 }